requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
scikit-learn>=0.24.2
Pillow>=8.3.1
tkinter 